safari-600.7-branch/LayoutTests/fast/dom/scroll-element-to-rect-expected.txt

Scrolled element to rect
PASS rect.left is computedLeft
PASS rect.top is computedTop
PASS successfullyParsed is true

TEST COMPLETE
